장소 검색 API - 검색 결과 수 5개 미만일 경우 다른 서비스에서 호출 하도록 수정 Swagger 적용 검색 이력 API 개발 로직 수정으로 인한 테스트 수정

+++ src/main/kotlin/com/place/search/presentation/external/request/KakaoLocalSearchRequest.kt
@@ -6,9 +6,9 @@ data class KakaoLocalSearchRequest(
 
     // 결과 페이지 번호
     // (최소: 1, 최대: 45, 기본대값: 1)
-    val page: Long? = 5,
+    val page: Int? = 5,
 
     // 한 페이지에 보여질 문서의 개수
     // (최소: 1, 최대: 45, 기본값: 15)
-    val size: Long? = 1,
+    val size: Int? = 1,
 )
